+++ apk/Changelog.md
@@ -9,6 +9,12 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Version November 27 2024 Apk Version 3.0.372
+==============
+
+* Update player dependencies to latest version
+* General improvements
+
 Version November 15 2024
 ==============